refactor: Custom Page Extensionsの導入と、ページのビューとロジックの分離を行なって。
