+++ app/src/pages/category/[slug]/[id].page.tsx
@@ -1,30 +1,21 @@
-import { NextPage, GetStaticPaths, GetStaticProps } from 'next';
+import { NextPage } from 'next';
 import Link from 'next/link';
 import React from 'react';
-
 import Breadcrumb from '@/components/Breadcrumb';
 import Layout from '@/components/Layout';
 import Paginate from '@/components/Paginate';
 import PostThumbnail from '@/components/PostThumbnail';
 import Seo from '@/components/Seo';
-
 import { config } from '@/config/app';
-
-import { fetchArticlesPage } from '@/lib/fetch-articles-page';
-import { fetchTaxonomyPage } from '@/lib/fetch-taxonomy-page';
-
 import { Content } from '@/types/content';
-import { TaxonomyIdsAndRelatedPosts } from '@/types/taxonomy';
-
 import { BlogCard, PostInfo } from '@/share/BlogCard';
 import { CategoryLabel } from '@/share/CategoryLabel';
 import { H2, H3 } from '@/share/Heading';
 import { TagArea } from '@/share/TagArea';
 import { TagLabel } from '@/share/TagLabel';
 import { TimeStamp } from '@/share/TimeStamp';
-
 import { dateFormat } from '@/utils/date-format';
-import { generateBuildPaginatePath } from '@/utils/generate-build-paginate-path';
+import { getStaticPaths, getStaticProps } from './index.hook';
 
 type Props = {
   contents: Content[];
@@ -32,8 +23,6 @@ type Props = {
   categorySlug: string;
   totalCount: number;
 };
-
-const paginateNum = config.paginateNum;
 
 const CategoryPage: NextPage<Props> = ({ contents, categoryName, categorySlug, totalCount }) => {
   const { siteTitle } = config.siteInfo;
@@ -84,43 +73,6 @@ const CategoryPage: NextPage<Props> = ({ contents, categoryName, categorySlug, t
   );
 };
 
-export const getStaticPaths: GetStaticPaths = async () => {
-  const data = await fetchTaxonomyPage<TaxonomyIdsAndRelatedPosts>('category', 'id,posts.id');
-  const results = generateBuildPaginatePath(data.contents);
-
-  const paths = results.map((path) => ({
-    params: {
-      slug: path.params.slug,
-      id: path.params.id.toString(),
-    },
-  }));
-
-  return {
-    paths,
-    fallback: false,
-  };
-};
-
-export const getStaticProps: GetStaticProps = async (context) => {
-  const slug = context?.params?.slug;
-  const id = context?.params?.id as string;
-  const offset = parseInt(id) * paginateNum - paginateNum;
-
-  const data = await fetchArticlesPage(offset, paginateNum, `category_field[equals]${slug}`);
-
-  const contents = data.contents;
-  const categoryName = contents[0].category_field.name;
-  const categorySlug = contents[0].category_field.id;
-  const totalCount = data.totalCount;
-
-  return {
-    props: {
-      contents,
-      categoryName,
-      categorySlug,
-      totalCount,
-    },
-  };
-};
+export { getStaticPaths, getStaticProps };
 
 export default CategoryPage;
