+++ app/src/pages/page/[id].page.tsx
